Create 'latest' symlink(s) when materializing schemas
Add option that if true, create's both latest.(json|yaml) as well as an
extensionless 'latest' symlink (if configured to do so) when materializing
schemas.  This symlinks will only ever be updated if jsonschema-tools
matereializes a schema that has a greater version than the version
of the schema that the latest symlink currently points at.

Bug: T245859