Fixes for structure tests when schema title does not match relative directory
The schema repository tests are dynamic.  The structure tests are grouped
by schema_title.  If something has gone wrong, and there is a materialized
schema file that has a title that does not match the directory it is in,
furhter test declarations will fail.

Prevent further test declarations if any schema titles and directories
are mismatched.

Bump to 0.7.7.