Add option to enforce bounds for number fields
This option configures a pair of numeric bounds that, at the
schema materialization level, will apply those bounds to number
fields (float, integer or number) by traversing the schema.

The change includes a small refactor that separates the logic that
modifies the materialized schema from the main function that
outputs the materialized schema to a path.

Bug: T258659