Refactor tests to allow selectively skipping certain test cases
Not all schemas are perfect.  There are times when schema rules change,
and old/legacy schemas don't conform to the new rules.

This adds a `skipSchemaTestCases` config option to allow
for matching schema $ids to test case names that should be skipped
for that schema.

This is currently only implemented for robustness tests, but
more should be implemented in a future patch.

Bug: T285006