When resolving allOf examples, only keep one examples list
This is cool, but I had some errors when actually merging examples like
this in real schema repos.  Perhaps something changed in
json-schema-merge-allof, but it now required me to add a defaultResolver
for non JSONSchema keywords that end up being passed to its mergeSchemas
function.  This could work, but I think the right thing to do is just
to keep the root schemas examples above all else.  This avoids
having to recurse into the examples objects themselves and merge them.

This means that examples in $ref-ed fragment schemas will
not be used, which I think is a more sensible behavior.

Bug: T270134