Add jsonschema-tools diff subcommand (#51)
This finds materialized schema versions that have been modified since
the previous commit, and then prints a diff between them.

This will make reviewing schema changes easier.  We can schedule
jsonschema-tools diff to be run as a CI job, and reviewers can
look at the output to compare e.g. version 1.0.0 to version 1.1.0, and so on.

In WMF CI, we should use dyff as the --diff-command to show semantic yaml diffs,
instead of textual diffs.

Bug: T321850

---------

Co-authored-by: Thomas Chin <tchin25@gmail.com>